c58

+++ es-app/src/guis/GuiMenu.cpp
@@ -4352,7 +4352,7 @@ void GuiMenu::deleteBtnJoyCfg(Window *mWindow, GuiSettings *systemConfiguration,
 		int err=0;
 		try {
 			OptionListComponent<std::string>* olc = dynamic_cast<OptionListComponent<std::string>*>(gc);
-			int **adress = &olc;
+			long unsigned int **adress = &olc;
 			SystemConf::getInstance()->set(prefixName + ".adress", std::to_string(adress));
 			if (olc == nullptr)
 				err=1;
